Add YOLOX support (#2644)
* change eval to evaluate

* fix merge

* fix merge

* change cascade_rcnn_x101 to a smaller centernet in detection test

* increase timeout from 15 to 120 for mim install

* update centernet version

* fix

* change to maskrcnn

* finished ft, testing val metric

* complete val metric integration for obj detection

* fix val bug

* fix val_metric error

* yolo voc finetune success

* refactors

* remove COLLECTION

* refactors

* refactor

* refactor

* black

* small fic

* small fix

* black

* lint

* lint

* lint

* fix pred without label

* fix ocr

* add comments

* add comments

* fix merge

* revert detection inference example change

* fix a conflict in matcher and update a dependency for mmocr

* decrease mim timeout to 60 second

* import mmcv only when use

* import mmcv only when use

* black

* black

* bug fix

* bug fix

* update detection eval example

* add todo mark

* cleaner code and black

* resolve several comments

* change infer_label_column_type_by's name

* fix typo

* add todo mark

* change constant map to mAP

* change mmcv_model constant to mmlab_models

* lint

* add todos and add if det to reset_index

* multi gpu training

* black

* add val_metric option and fix some TODOs

* small refactors

* black

* lint

* fix circular import

* black

* small fix

* fix collate and better integrate with lightning for detection

* remove unnecessary save/load

* remove train_to_val and val_to_train in litmmdet

* fix map

* temp fix for MMOCR, need to update

* enable multigpu inference

* merge

* lint

* fix single gpu inference

* add doc

* skip ocr test for now

* black

* black

* rename output_shape to num_classes

* fix

* add a TODO

* add train from scratch for detection model

* fix hardcoding for voc and coco while load coco format data

* black

* small fix and add voc2coco tool

* update voc2coco.py

* fix type

* fix voc2coco

* fix

* fix classes hardcode and add support for two stage lr

* fix single gpu eval and fix torchmetrics eval

* add support for voc format input

* fix a word

* add preset, fix eval bug for voc

* some refactors

* black and small refactor

* lint

* fix assets loading

* fix

* black

* add docs

* open PIL safely to avoid resource warning

* add fasterrcnn 2 stage lr and black

* add set_num_gpus

* add evaluation tutorial

* refine documents

* update ddetr

* fix typo

* change ref names

* merge

* add eval on voc and default two stage setting

* finish tutorial in datasets preparation

* small fix

* update tutorial and codes

* add EOF line and remove finetune script

* black

* fix typo in docs

* fix wordings and add get_detection_classes, fix voc2coco, fix predict

* fix type

* small update

* fix predict df error

* fix voc2coco

* add finetune tutorial and support vfnet, raise error for mask models

* add voc2coco tutorial

* add eof

* update doc

* minor

* support more models and skip image size if not in config

* black

* update docs

* add load predictor and change to problem type

* update fps information

* update load preditor example

* fix type

* update tutorial

* move val_metric

* add presets and fix val_metric in tuto and example

* black

* remove mypy in test_common

* fix

* change preset name

* restart ci

* index tutorial and fix save_path problem

* fix get_num_gpu

* fix save_path

* fix tuto and example typo and path error

* add a TODO and force detection not realtime in _predict

* fix voc2coco

* black

* bug bash + quick start

* merge

* merge

* fix save_path problem for DDP

* black

* update warning

* update tutorial, lint

* update tuto

* update voc to coco

* lint

* refine index

* lint

* update docs in multimodal

* update the multimodal index

* add doc depth

* fix save path init

* restart ci

* restart ci

* fix save_path in fit

* fix save_path

* fix save_path

* update quick start

* update cli in doc

* change prefix coco to postfix cocoformat for voc2coco

* fix s3 path and index error

* restart ci

* install necessary packages

* restart ci

* add build mmcv in .github

* fix mmlab installation

* fix

* fix mmcv install in tuto

* fix

* fix

* fix index

* update docs

* refine quick start

* remove unnecessary warnings

* restart ci

* replace voc2012 to 2007

* fix tuto and cli tool

* merge

* fix docs

* change max depth

* add all detection tuto in multimodal index

* restart ci

* fix max depth

* fix installation error for pycocotools

* remove windows

* add benchmarking

* benchmarking

* fix voc2coco

* update for new machine

* add o365 download script

* create save_path in init and remove warning in fit

* remove unecessary line

* fix

* +yoloX with hardcode

* update

* remove scripts

* remove files

* fix typo

* add mmdet configs in repo

* remove configs in example

* remove hardcoding

* black

* lint

* fix circular import

* fix

* nit

* move clean_ckpts to fit

* renaming

* black

+++ multimodal/src/autogluon/multimodal/predictor.py
@@ -529,6 +529,7 @@ class MultiModalPredictor:
         presets: Optional[str] = None,
         config: Optional[dict] = None,
         tuning_data: Optional[Union[pd.DataFrame, str]] = None,
+        max_num_tuning_data: Optional[int] = None,
         id_mappings: Optional[Union[Dict[str, Dict], Dict[str, pd.Series]]] = None,
         time_limit: Optional[int] = None,
         save_path: Optional[str] = None,
@@ -539,6 +540,7 @@ class MultiModalPredictor:
         seed: Optional[int] = 123,
         standalone: Optional[bool] = True,
         hyperparameter_tune_kwargs: Optional[dict] = None,
+        clean_ckpts: Optional[bool] = True,
     ):
         """
         Fit MultiModalPredictor predict label column of a dataframe based on the other columns,
@@ -643,6 +645,8 @@ class MultiModalPredictor:
                     scheduler_init_args: Optional[dict] = None
                         If provided str to `searcher`, you can optionally provide custom init_args to the searcher
                         You don't need to worry about `metric` and `mode`. AutoGluon will figure it out by itself.
+        clean_ckpts
+            Whether to clean the checkpoints of each validation step after training.
 
         Returns
         -------
@@ -678,7 +682,13 @@ class MultiModalPredictor:
             self._detection_anno_train = train_data
             train_data = from_coco_or_voc(train_data, "train")
             if tuning_data is not None:
+                self.detection_anno_train = tuning_data
                 tuning_data = from_coco_or_voc(tuning_data, "val")
+                if max_num_tuning_data is not None:
+                    if len(tuning_data) > max_num_tuning_data:
+                        tuning_data = tuning_data.sample(
+                            n=max_num_tuning_data, replace=False, random_state=seed
+                        ).reset_index(drop=True)
 
         if hyperparameter_tune_kwargs is not None:
             # TODO: can we support hyperparameters being the same format as regular training?
@@ -821,6 +831,7 @@ class MultiModalPredictor:
             teacher_predictor=teacher_predictor,
             standalone=standalone,
             hpo_mode=(hyperparameter_tune_kwargs is not None),  # skip average checkpoint if in hpo mode
+            clean_ckpts=clean_ckpts,
         )
 
         if hyperparameter_tune_kwargs is not None:
@@ -994,6 +1005,7 @@ class MultiModalPredictor:
         teacher_predictor: Union[str, MultiModalPredictor] = None,
         hpo_mode: bool = False,
         standalone: bool = True,
+        clean_ckpts: bool = True,
         **hpo_kwargs,
     ):
 
@@ -1116,6 +1128,7 @@ class MultiModalPredictor:
                 validation_metric_name=validation_metric_name,
                 strict_loading=not trainable_param_names,
                 standalone=standalone,
+                clean_ckpts=clean_ckpts,
             )
 
             return self
@@ -1409,6 +1422,7 @@ class MultiModalPredictor:
                     strategy=strategy,
                     strict_loading=not trainable_param_names,  # Not strict loading if using parameter-efficient finetuning
                     standalone=standalone,
+                    clean_ckpts=clean_ckpts,
                 )
         else:
             sys.exit(f"Training finished, exit the process with global_rank={trainer.global_rank}...")
@@ -1426,6 +1440,7 @@ class MultiModalPredictor:
         last_ckpt_path=None,
         strict_loading=True,
         standalone=True,
+        clean_ckpts=True,
     ):
         # FIXME: we need to change validation_metric to evaluation_metric for model choosing
         # since we called self.evaluate. Below is a temporal fix for NER.
@@ -1552,16 +1567,17 @@ class MultiModalPredictor:
 
         torch.save(checkpoint, os.path.join(save_path, MODEL_CHECKPOINT))
 
-        # clean old checkpoints + the intermediate files stored
-        for per_path in top_k_model_paths:
-            if os.path.isfile(per_path):
-                os.remove(per_path)
-        # remove the yaml file after cleaning the checkpoints
-        if os.path.isfile(best_k_models_yaml_path):
-            os.remove(best_k_models_yaml_path)
-        # clean the last checkpoint
-        if os.path.isfile(last_ckpt_path):
-            os.remove(last_ckpt_path)
+        if clean_ckpts:
+            # clean old checkpoints + the intermediate files stored
+            for per_path in top_k_model_paths:
+                if os.path.isfile(per_path):
+                    os.remove(per_path)
+            # remove the yaml file after cleaning the checkpoints
+            if os.path.isfile(best_k_models_yaml_path):
+                os.remove(best_k_models_yaml_path)
+            # clean the last checkpoint
+            if os.path.isfile(last_ckpt_path):
+                os.remove(last_ckpt_path)
 
     def _default_predict(
         self,
